handle missing intel variable.

+++ src/@ionic-native/plugins/intel-security/index.ts
@@ -1,6 +1,7 @@
 import { Plugin } from '@ionic-native/core';
 import { Injectable } from '@angular/core';
 
+declare var window: any;
 declare var intel: any;
 
 /**
@@ -27,14 +28,15 @@ declare var intel: any;
 @Injectable()
 export class IntelSecurity {
 
-  public storage;
-  public data;
+  public storage: IntelSecurityStorage;
+  public data: IntelSecurityData;
 
   constructor() {
     this.storage = new IntelSecurityStorage();
     this.data = new IntelSecurityData();
+
     try {
-      if (!intel)
+      if (!window.intel)
         throw true;
     } catch (e) {
       console.warn('Native: Intel App Security API is not installed or you are running on a browser.');
@@ -53,11 +55,10 @@ export class IntelSecurityData {
   * @returns {Promise<any>} Returns a Promise that resolves with the instanceID of the created data instance, or rejects with an error.
   */
   createFromData(data: string): Promise<any> {
-    try {
-      return intel.security.secureData.createFromData({ data: data });
-    } catch (e) {
-      console.warn('Native: tried calling IntelSecurity.data.createFromData, but Intel App Security API is not available. Make sure to run in a device/simulator.');
+    if (!window.intel) {
+      return Promise.reject(console.warn('Native: tried calling IntelSecurity.data.createFromData, but Intel App Security API is not available. Make sure to run in a device/simulator.'));
     }
+    return intel.security.secureData.createFromData({ data: data });
   }
 
   /**
@@ -66,11 +67,10 @@ export class IntelSecurityData {
    * @returns {Promise<any>} Returns a Promise that resolves with the instanceID of the created data instance, or rejects with an error.
    */
   createFromSealedData(sealedData: string): Promise<any> {
-    try {
-      return intel.security.secureData.createFromSealedData({ sealedData: sealedData });
-    } catch (e) {
-      console.warn('Native: tried calling IntelSecurity.data.createFromSealedData, but Intel App Security API is not available. Make sure to run in a device/simulator.');
+    if (!window.intel) {
+      return Promise.reject(console.warn('Native: tried calling IntelSecurity.data.createFromSealedData, but Intel App Security API is not available. Make sure to run in a device/simulator.'));
     }
+    return intel.security.secureData.createFromSealedData({ sealedData: sealedData });
   }
 
   /**
@@ -79,12 +79,10 @@ export class IntelSecurityData {
    * @returns {Promise<string>} Returns a Promise that resolves to the data as plain-text, or rejects with an error.
    */
   getData(instanceID: any): Promise<string> {
-    try {
-      return intel.security.secureData.getData(instanceID);
-
-    } catch (e) {
-      console.warn('Native: tried calling IntelSecurity.data.getData, but Intel App Security API is not available. Make sure to run in a device/simulator.');
+    if (!window.intel) {
+      return Promise.reject(console.warn('Native: tried calling IntelSecurity.data.getData, but Intel App Security API is not available. Make sure to run in a device/simulator.'));
     }
+    return intel.security.secureData.getData(instanceID);
   }
 
   /**
@@ -93,11 +91,10 @@ export class IntelSecurityData {
    * @returns {Promise<any>} Returns a Promise that resolves to the sealed data, or rejects with an error.
    */
   getSealedData(instanceID: any): Promise<any> {
-    try {
-      return intel.security.secureData.getSealedData(instanceID);
-    } catch (e) {
-      console.warn('Native: tried calling IntelSecurity.data.getSealedData, but Intel App Security API is not available. Make sure to run in a device/simulator.');
+    if (!window.intel) {
+      return Promise.reject(console.warn('Native: tried calling IntelSecurity.data.getSealedData, but Intel App Security API is not available. Make sure to run in a device/simulator.'));
     }
+    return intel.security.secureData.getSealedData(instanceID);
   }
 
   /**
@@ -106,11 +103,10 @@ export class IntelSecurityData {
    * @returns {Promise<string>} Returns a Promise that resolves to the tag, or rejects with an error.
    */
   getTag(instanceID: any): Promise<string> {
-     try {
-      return intel.security.secureData.getTag(instanceID);
-    } catch (e) {
-      console.warn('Native: tried calling IntelSecurity.data.getTag, but Intel App Security API is not available. Make sure to run in a device/simulator.');
+    if (!window.intel) {
+      return Promise.reject(console.warn('Native: tried calling IntelSecurity.data.getTag, but Intel App Security API is not available. Make sure to run in a device/simulator.'));
     }
+    return intel.security.secureData.getTag(instanceID);
   }
 
   /**
@@ -119,11 +115,10 @@ export class IntelSecurityData {
    * @returns {Promise<any> Returns a promise that resolves to the policy object, or rejects with an error.
    */
   getPolicy(instanceID: any): Promise<any> {
-     try {
-      return intel.security.secureData.getPolicy(instanceID);
-    } catch (e) {
-      console.warn('Native: tried calling IntelSecurity.data.getPolicy, but Intel App Security API is not available. Make sure to run in a device/simulator.');
+    if (!window.intel) {
+      return Promise.reject(console.warn('Native: tried calling IntelSecurity.data.getPolicy, but Intel App Security API is not available. Make sure to run in a device/simulator.'));
     }
+    return intel.security.secureData.getPolicy(instanceID);
   }
 
   /**
@@ -132,11 +127,10 @@ export class IntelSecurityData {
    * @returns {Promise<Array>} Returns a promise that resolves to an array of owners' unique IDs, or rejects with an error.
    */
   getOwners(instanceID: any): Promise<Array<any>> {
-     try {
-      return intel.security.secureData.getOwners(instanceID);
-    } catch (e) {
-      console.warn('Native: tried calling IntelSecurity.data.getOwners, but Intel App Security API is not available. Make sure to run in a device/simulator.');
+    if (!window.intel) {
+      return Promise.reject(console.warn('Native: tried calling IntelSecurity.data.getOwners, but Intel App Security API is not available. Make sure to run in a device/simulator.'));
     }
+    return intel.security.secureData.getOwners(instanceID);
   }
 
   /**
@@ -145,11 +139,10 @@ export class IntelSecurityData {
    * @returns {Promise<Number>} Returns a promsie that resolves to the creator's unique ID, or rejects with an error.
    */
   getCreator(instanceID: any): Promise<Number> {
-     try {
-      return intel.security.secureData.getCreator(instanceID);
-    } catch (e) {
-      console.warn('Native: tried calling IntelSecurity.data.getCreator, but Intel App Security API is not available. Make sure to run in a device/simulator.');
+    if (!window.intel) {
+      return Promise.reject(console.warn('Native: tried calling IntelSecurity.data.getCreator, but Intel App Security API is not available. Make sure to run in a device/simulator.'));
     }
+    return intel.security.secureData.getCreator(instanceID);
   }
 
   /**
@@ -158,11 +151,10 @@ export class IntelSecurityData {
    * @returns {Promise<Array>} Returns a promise that resolves to a list of web owners, or rejects with an error.
    */
   getWebOwners(instanceID: any): Promise<Array<any>> {
-     try {
-      return intel.security.secureData.getWebOwners(instanceID);
-    } catch (e) {
-      console.warn('Native: tried calling IntelSecurity.data.getWebOwners, but Intel App Security API is not available. Make sure to run in a device/simulator.');
+    if (!window.intel) {
+      return Promise.reject(console.warn('Native: tried calling IntelSecurity.data.getWebOwners, but Intel App Security API is not available. Make sure to run in a device/simulator.'));
     }
+    return intel.security.secureData.getWebOwners(instanceID);
   }
 
   /**
@@ -172,11 +164,10 @@ export class IntelSecurityData {
    * @returns {Promise<any>} Returns a promise that resolves with no parameters, or rejects with an error.
    */
   changeExtraKey(instanceID: any, extraKey: Number): Promise<any> {
-     try {
-      return intel.security.secureData.changeExtraKey({ instanceID: instanceID, extraKey: extraKey });
-    } catch (e) {
-      console.warn('Native: tried calling IntelSecurity.data.changeExtraKey, but Intel App Security API is not available. Make sure to run in a device/simulator.');
+    if (!window.intel) {
+      return Promise.reject(console.warn('Native: tried calling IntelSecurity.data.changeExtraKey, but Intel App Security API is not available. Make sure to run in a device/simulator.'));
     }
+    return intel.security.secureData.changeExtraKey({ instanceID: instanceID, extraKey: extraKey });
   }
 
   /**
@@ -185,11 +176,10 @@ export class IntelSecurityData {
    * @returns {Promise<any>} Returns a promise that resovles with no parameters, or rejects with an error.
    */
   destroy(instanceID: any): Promise<any> {
-     try {
-      return intel.security.secureData.destroy(instanceID);
-    } catch (e) {
-      console.warn('Native: tried calling IntelSecurity.data.destroy, but Intel App Security API is not available. Make sure to run in a device/simulator.');
+    if (!window.intel) {
+      return Promise.reject(console.warn('Native: tried calling IntelSecurity.data.destroy, but Intel App Security API is not available. Make sure to run in a device/simulator.'));
     }
+    return intel.security.secureData.destroy(instanceID);
   }
 
 }
@@ -205,11 +195,10 @@ export class IntelSecurityStorage {
    * @returns {Promise<any>} Returns a Promise that resolves with no parameters, or rejects with an error.
    */
   delete(id: string): Promise<any> {
-     try {
-      return intel.security.secureStorage.delete({ id: id });
-    } catch (e) {
-      console.warn('Native: tried calling IntelSecurity.storage.delete, but Intel App Security API is not available. Make sure to run in a device/simulator.');
+    if (!window.intel) {
+      return Promise.reject(console.warn('Native: tried calling IntelSecurity.storage.delete, but Intel App Security API is not available. Make sure to run in a device/simulator.'));
     }
+    return intel.security.secureStorage.delete({ id: id });
   }
 
   /**
@@ -218,11 +207,10 @@ export class IntelSecurityStorage {
    * @returns {Promise<string>} Returns a Promise that resolved with the data as plain-text, or rejects with an error.
    */
   read(id: string): Promise<string> {
-     try {
-      return intel.security.secureStorage.read({ id: id });
-    } catch (e) {
-      console.warn('Native: tried calling IntelSecurity.storage.read, but Intel App Security API is not available. Make sure to run in a device/simulator.');
+    if (!window.intel) {
+      return Promise.reject(console.warn('Native: tried calling IntelSecurity.storage.read, but Intel App Security API is not available. Make sure to run in a device/simulator.'));
     }
+    return intel.security.secureStorage.read({ id: id });
   }
 
   /**
@@ -231,11 +219,10 @@ export class IntelSecurityStorage {
    * @param instanceID {number} Valid secure data instance ID
    */
   write(id: string, instanceID: number): Promise<any> {
-     try {
-      return intel.security.secureStorage.write({ id: id, instanceID: instanceID });
-    } catch (e) {
-      console.warn('Native: tried calling IntelSecurity.storage.write, but Intel App Security API is not available. Make sure to run in a device/simulator.');
+    if (!window.intel) {
+      return Promise.reject(console.warn('Native: tried calling IntelSecurity.storage.write, but Intel App Security API is not available. Make sure to run in a device/simulator.'));
     }
+    return intel.security.secureStorage.write({ id: id, instanceID: instanceID });
   }
 
 }
